sql/expression: split functions and aggregations
