+++ sql/expression/common.go
@@ -1,6 +1,8 @@
 package expression
 
-import "gopkg.in/src-d/go-mysql-server.v0/sql"
+import (
+	"gopkg.in/src-d/go-mysql-server.v0/sql"
+)
 
 // UnaryExpression is an expression that has only one children.
 type UnaryExpression struct {
@@ -31,37 +33,4 @@ func (p BinaryExpression) Resolved() bool {
 // IsNullable returns whether the expression can be null.
 func (p BinaryExpression) IsNullable() bool {
 	return p.Left.IsNullable() || p.Right.IsNullable()
-}
-
-var defaultFunctions = map[string]sql.Function{
-	"count": sql.Function1(func(e sql.Expression) sql.Expression {
-		return NewCount(e)
-	}),
-	"min": sql.Function1(func(e sql.Expression) sql.Expression {
-		return NewMin(e)
-	}),
-	"max": sql.Function1(func(e sql.Expression) sql.Expression {
-		return NewMax(e)
-	}),
-	"avg": sql.Function1(func(e sql.Expression) sql.Expression {
-		return NewAvg(e)
-	}),
-	"is_binary": sql.Function1(NewIsBinary),
-	"substring": sql.FunctionN(NewSubstring),
-	"year":      sql.Function1(NewYear),
-	"month":     sql.Function1(NewMonth),
-	"day":       sql.Function1(NewDay),
-	"hour":      sql.Function1(NewHour),
-	"minute":    sql.Function1(NewMinute),
-	"second":    sql.Function1(NewSecond),
-	"dayofyear": sql.Function1(NewDayOfYear),
-}
-
-// RegisterDefaults registers the aggregations in the catalog.
-func RegisterDefaults(c *sql.Catalog) error {
-	for k, v := range defaultFunctions {
-		c.RegisterFunction(k, v)
-	}
-
-	return nil
 }
