+++ sql/expression/function/substring_test.go
@@ -1,4 +1,4 @@
-package expression
+package function
 
 import (
 	"context"
@@ -6,61 +6,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"gopkg.in/src-d/go-mysql-server.v0/sql"
+	"gopkg.in/src-d/go-mysql-server.v0/sql/expression"
 )
-
-func TestIsBinary(t *testing.T) {
-	f := NewIsBinary(NewGetField(0, sql.Blob, "blob", true))
-
-	testCases := []struct {
-		name     string
-		row      sql.Row
-		expected bool
-	}{
-		{"binary", sql.NewRow([]byte{0, 1, 2}), true},
-		{"not binary", sql.NewRow([]byte{1, 2, 3}), false},
-		{"null", sql.NewRow(nil), false},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			require.Equal(t, tt.expected, eval(t, f, tt.row))
-		})
-	}
-}
-
-func TestSubstringArity(t *testing.T) {
-	expr := NewGetField(0, sql.Int64, "foo", false)
-	testCases := []struct {
-		name string
-		args []sql.Expression
-		ok   bool
-	}{
-		{"0 args", nil, false},
-		{"1 args", []sql.Expression{expr}, false},
-		{"2 args", []sql.Expression{expr, expr}, true},
-		{"3 args", []sql.Expression{expr, expr, expr}, true},
-		{"4 args", []sql.Expression{expr, expr, expr, expr}, false},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-			f, err := NewSubstring(tt.args...)
-			if tt.ok {
-				require.NotNil(f)
-				require.NoError(err)
-			} else {
-				require.Error(err)
-			}
-		})
-	}
-}
 
 func TestSubstring(t *testing.T) {
 	f, err := NewSubstring(
-		NewGetField(0, sql.Text, "str", true),
-		NewGetField(1, sql.Int32, "start", false),
-		NewGetField(2, sql.Int64, "len", false),
+		expression.NewGetField(0, sql.Text, "str", true),
+		expression.NewGetField(1, sql.Int32, "start", false),
+		expression.NewGetField(2, sql.Int64, "len", false),
 	)
 	require.NoError(t, err)
